fix: navigation for settings and space
